refactor(robot-server): Single-source filenames (#15985)
