+++ robot-server/robot_server/data_files/dependencies.py
@@ -1,7 +1,6 @@
 """FastAPI dependencies for data files endpoints."""
 from pathlib import Path
 from asyncio import Lock as AsyncLock
-from typing import Final
 from anyio import Path as AsyncPath
 
 from fastapi import Depends
@@ -17,12 +16,11 @@ from robot_server.persistence.fastapi_dependencies import (
     get_active_persistence_directory,
     get_sql_engine,
 )
+from robot_server.persistence.file_and_directory_names import DATA_FILES_DIRECTORY
 from robot_server.deletion_planner import DataFileDeletionPlanner
 from .data_files_store import DataFilesStore
 from .file_auto_deleter import DataFileAutoDeleter
 
-
-_DATA_FILES_SUBDIRECTORY: Final = "data_files"
 
 _data_files_directory_init_lock = AsyncLock()
 _data_files_directory_accessor = AppStateAccessor[Path]("data_files_directory")
@@ -39,7 +37,7 @@ async def get_data_files_directory(
     async with _data_files_directory_init_lock:
         data_files_dir = _data_files_directory_accessor.get_from(app_state)
         if data_files_dir is None:
-            data_files_dir = persistent_directory / _DATA_FILES_SUBDIRECTORY
+            data_files_dir = persistent_directory / DATA_FILES_DIRECTORY
             await AsyncPath(data_files_dir).mkdir(exist_ok=True)
             _data_files_directory_accessor.set_on(app_state, data_files_dir)
