+++ robot-server/robot_server/persistence/__init__.py
@@ -1,5 +1,1 @@
 """Support for persisting data across device reboots."""
-
-from ._files_and_directories import LATEST_VERSION_DIRECTORY
-
-__all__ = ["LATEST_VERSION_DIRECTORY"]
