--- robot-server/robot_server/persistence/_files_and_directories.py
@@ -1,8 +1,0 @@
-"""Files and directories in persistent storage."""
-from typing import Final
-
-DECK_CONFIGURATION_FILE: Final = "deck_configuration.json"
-PROTOCOLS_DIRECTORY: Final = "protocols"
-DATA_FILES_DIRECTORY: Final = "data_files"
-DB_FILE: Final = "robot_server.db"
-LATEST_VERSION_DIRECTORY: Final = "6"
